fix predict
