+++ deploy/lite/benchmark/benchmark.sh
